rename theory to dataloop and getProducts to product for readability
