+++ ui/spec/regression.test/shareProducts.spec.ts
@@ -1,8 +1,8 @@
 // Data
 import { env } from "../data/environment.dat";
-import { getProducts, IProduct } from "../data/test.dat";
+import { products, IProduct } from "../data/test.dat";
 
-import { using } from "../lib/theory";
+import { using } from "../lib/dataLoop";
 
 // Page Objects
 import {
@@ -14,7 +14,7 @@ const productPg = new ProductPg();
 fixture `AutoPrac Product Social Media Sharing`
 .page(env.url);
 
-using (getProducts(), (product: IProduct) => {
+using (products(), (product: IProduct) => {
 
     test("Verify social media sharing links for: " + product.productName, async (t) => {
         await t.maximizeWindow();
