+++ ui/spec/smoke.test/searchProducts.spec.ts
@@ -1,8 +1,8 @@
 // Data
 import { env } from "../data/environment.dat";
-import { getProducts, IProduct } from "../data/test.dat";
+import { products, IProduct } from "../data/test.dat";
 
-import { using } from "../lib/theory";
+import { using } from "../lib/dataLoop";
 
 // Page Objects
 import {
@@ -14,7 +14,7 @@ const productPg = new ProductPg();
 fixture `AutoPrac Product Searching`
 .page(env.url);
 
-using (getProducts(), (product: IProduct) => {
+using (products(), (product: IProduct) => {
 
     test("Search for item: " + product.productName, async (t) => {
         await t
